basic functionalities=ok

+++ app/Http/Controllers/CategoryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Product;
 use App\Models\Category;
 use Illuminate\Http\Request;
 
@@ -14,9 +15,25 @@ class CategoryController extends Controller
 
     public function index()
     {
-        $categories = Category::with("user")->simplePaginate(20);
+        $categories = Category::latest()
+            ->with("user")
+            ->simplePaginate(20);
 
         return view("category.index", compact("categories"));
+    }
+
+    public function show(Category $category)
+    {
+        $products = Product::latest()
+            ->where("category_id", $category->id)
+            ->simplePaginate(10);
+
+        $product_count = Product::where("category_id", $category->id)->count();
+
+        return view(
+            "category.show",
+            compact(["category", "products", "product_count"])
+        );
     }
 
     public function create()
@@ -47,6 +64,10 @@ class CategoryController extends Controller
 
     public function update(Request $request, Category $category)
     {
+        if ($category->user_id != auth()->user()->id) {
+            return back();
+        }
+
         $request->validate([
             "name" => ["required", "min:3", "max:100", "unique:users"],
             "description" => ["required", "min:3"],
@@ -62,6 +83,10 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        if ($category->user_id != auth()->user()->id) {
+            return back();
+        }
+
         $category->delete();
 
         return redirect()->route("category.index");
